feat(ibkr-fractional): exec layer float-capable + IBKR equities trade fractional
The sub-$35 cap was a code artifact, not just a broker rule: RH fractional can't
carry a stop (whole-share only), but IBKR supports fractional + stop — yet the
exec/reconcile stack was integer-only (built for futures), so the IBKR lane also
did int(size/c) and skipped <1-share. Now:
- hardening/exec_manager.py: TradeIntent.qty + ExecutionResult.filled_qty int->float
  (backward-compatible: futures still pass integer contracts).
- bot/live_equities_ibkr.py: fractional sizing (size_usd/c, no >=1-share gate).
  IBKR paper lane now trades the FULL 1,459-name universe fractional w/ native stop.
257 tests pass.

+++ hardening/exec_manager.py
@@ -63,7 +63,7 @@ class TradeIntent:
     symbol: str         # 'MES'
     action: str         # 'BUY' / 'SELL'
     side: str           # 'LONG' / 'SHORT'
-    qty: int
+    qty: float
     order_type: str     # 'MKT' / 'STP'
     stop_price: float   # 0.0 when no protective stop
     contract_month: str
@@ -117,7 +117,7 @@ class IntentStore:
 @dataclass
 class ExecutionResult:
     status: str          # FILLED | PARTIAL | REJECTED | UNKNOWN | DUPLICATE
-    filled_qty: int = 0
+    filled_qty: float = 0.0
     avg_px: float = 0.0
     signal_id: str = ''
     intent_id: str = ''
